Added protected routes for goals

+++ backend/controllers/goalController.js
@@ -1,14 +1,15 @@
 const asyncHandler = require("express-async-handler")
 const Goal = require("../models/goalModel")
+const User = require("../models/userModel")
 
 //@desc Get goals
 //@route GET /api/goals
 //access private
 const getGoals = asyncHandler(async (req, res) => {
-	const goals = await Goal.find()
+	const goals = await Goal.find({ user: req.user.id })
 	res.status(200).json(goals)
 })
-//@desc Set goals
+//@desc Create new goals
 //@route SET /api/goals
 //access private
 const setGoal = asyncHandler(async (req, res) => {
@@ -18,6 +19,7 @@ const setGoal = asyncHandler(async (req, res) => {
 	}
 	const goal = await Goal.create({
 		text: req.body.text,
+		user: req.user.id,
 	})
 	res.status(200).json(goal)
 })
@@ -30,6 +32,20 @@ const updateGoal = asyncHandler(async (req, res) => {
 	if (!goal) {
 		res.status(400)
 		throw new Error("Goal not found")
+	}
+
+	const user = await User.findById(req.user.id)
+
+	//checkfor user
+	if (!user) {
+		res.status(401)
+		throw new Error("User not found")
+	}
+
+	//make sure the logged in user matches the goal user
+	if (goal.user.toString() !== user.id) {
+		res.status(401)
+		throw new Error("User not authorized")
 	}
 
 	const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, { new: true })
@@ -45,6 +61,20 @@ const deleteGoal = asyncHandler(async (req, res) => {
 	if (!goal) {
 		res.status(400)
 		throw new Error("Goal doesn't exist!")
+	}
+
+	const user = await User.findById(req.user.id)
+
+	//checkfor user
+	if (!user) {
+		res.status(401)
+		throw new Error("User not found")
+	}
+
+	//make sure the logged in user matches the goal user
+	if (goal.user.toString() !== user.id) {
+		res.status(401)
+		throw new Error("User not authorized")
 	}
 
 	await Goal.findByIdAndRemove(req.params.id)
